Changed float type.

+++ src/main/rust/erik/src/evaluation/ngram_evaluation.rs
@@ -1,7 +1,7 @@
 use super::ngram_table::NGramTable;
 
 /// Sums the values of the ngrams in the text. Useful as a fitness function.
-pub fn evaluate_ngrams<T: NGramTable + Sync>(text: &Vec<u8>, table: &T) -> f64 {
+pub fn evaluate_ngrams<T: NGramTable + Sync>(text: &Vec<u8>, table: &T) -> f32 {
     let size = table.get_ngram_size();
     text.windows(size)
         .map(|tetragram| table.get_value(tetragram))
